gemini_ocr_api_railway/requirements.txt
Flask==2.3.3
google-generativeai==0.5.2
Pillow==10.3.0